app/requirements.txt
flask==2.3.2
pymongo==4.3.3
openai==0.27.7
flask-jwt-extended==4.4.4
flask-cors==3.0.10
Werkzeug==2.3.4
markupsafe==2.1.2
google-auth==2.18.1
Flask-WTF==1.1.1
